+++ app/Mail/Verify.php
@@ -1,0 +1,55 @@
+<?php
+
+namespace App\Mail;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Mail\Mailable;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\URL;
+
+class Verify extends Mailable
+{
+    use Queueable, SerializesModels;
+
+    protected $user;
+
+    /**
+     * Create a new message instance.
+     *
+     * @return void
+     */
+    public function __construct($user)
+    {
+        $this->user = $user;
+    }
+
+    /**
+     * Build the message.
+     *
+     * @return $this
+     */
+    public function build()
+    {
+        return $this->markdown('emails.user.verify')->with([
+          'user' => $this->user,
+          'url' => $this->verificationUrl(),
+        ]);
+    }
+
+    /**
+     * Get the verification URL for the given notifiable.
+     *
+     * @return string
+     */
+    protected function verificationUrl()
+    {
+        return Url::temporarySignedRoute(
+          'verification.verify',
+          Carbon::now()->addMinutes(Config::get('auth.verification.expire', 60)),
+          ['id' => $this->user->getKey()]
+        );
+    }
+}